bugfix with long id in prepopulate

+++ src/main/java/de/eimantas/eimantasbackend/entities/Booking.java
@@ -20,7 +20,7 @@ public class Booking {
   private String name;
   private LocalDateTime startdate;
   private LocalDateTime endDate;
-  private int projectId;
+  private long projectId;
   @Id
   @GeneratedValue
   private long serverBookingId;
